Move pcg4d to test/utils.c
This commit introduces an internal module test/utils.c where we can
put test-specific utilities to prevent duplication. The pcg4d random
number generator that we use in the three test executables has been
moved into this module.
